+++ src/promptflow-devkit/tests/conftest.py
@@ -41,19 +41,20 @@ except ImportError:
     import time
 
     from promptflow._cli._pf._service import _start_background_service_on_unix, _start_background_service_on_windows
-    from promptflow._sdk._constants import PF_SERVICE_HOST
-    from promptflow._sdk._service.utils.utils import get_pfs_port
+    from promptflow._sdk._service.utils.utils import get_pfs_host, get_pfs_host_after_check_wildcard, get_pfs_port
 
-    def invoke_prompt_flow_service() -> str:
-        port = str(get_pfs_port())
+    def invoke_prompt_flow_service():
+        service_host = get_pfs_host()
+        host = get_pfs_host_after_check_wildcard(service_host)
+        port = str(get_pfs_port(host))
         if platform.system() == "Windows":
-            _start_background_service_on_windows(port)
+            _start_background_service_on_windows(port, service_host)
         else:
-            _start_background_service_on_unix(port)
+            _start_background_service_on_unix(port, service_host)
         time.sleep(20)
-        response = requests.get(f"http://{PF_SERVICE_HOST}:{port}/heartbeat")
+        response = requests.get(f"http://{host}:{port}/heartbeat")
         assert response.status_code == 200, "prompt flow service is not healthy via /heartbeat"
-        return port
+        return port, host
 
 
 load_dotenv()
@@ -325,9 +326,9 @@ def csharp_test_project_class_init_flex_flow() -> CSharpProject:
 @pytest.fixture(scope="session")
 def otlp_collector():
     """A session scope fixture, a separate standby prompt flow service serves as OTLP collector."""
-    port = invoke_prompt_flow_service()
+    port, service_host = invoke_prompt_flow_service()
     # mock invoke prompt flow service as it has been invoked already
-    with mock.patch("promptflow._sdk._tracing._invoke_pf_svc", return_value=port), mock.patch(
+    with mock.patch("promptflow._sdk._tracing._invoke_pf_svc", return_value=(port, service_host)), mock.patch(
         "promptflow._sdk._tracing.is_pfs_service_healthy", return_value=True
     ):
         yield