+++ src/promptflow-evals/promptflow/evals/evaluators/_relevance/_relevance.py
@@ -9,39 +9,59 @@ import numpy as np
 
 from promptflow.client import load_flow
 from promptflow.core import AzureOpenAIModelConfiguration
+try:
+    from ..._user_agent import USER_AGENT
+except ImportError:
+    USER_AGENT = None
 
 
 class RelevanceEvaluator:
+    """
+    Initialize a relevance evaluator configured for a specific Azure OpenAI model.
+
+    :param model_config: Configuration for the Azure OpenAI model.
+    :type model_config: AzureOpenAIModelConfiguration
+
+    **Usage**
+
+    .. code-block:: python
+
+        eval_fn = RelevanceEvaluator(model_config)
+        result = eval_fn(
+            question="What is the capital of Japan?",
+            answer="The capital of Japan is Tokyo.",
+            context="Tokyo is Japan's capital, known for its blend of traditional culture \
+                and technological advancements.")
+
+    **Output format**
+
+    .. code-block:: python
+
+        {
+            "gpt_relevance": 3.0
+        }
+    """
+
     def __init__(self, model_config: AzureOpenAIModelConfiguration):
-        """
-        Initialize an evaluator configured for a specific Azure OpenAI model.
-
-        :param model_config: Configuration for the Azure OpenAI model.
-        :type model_config: AzureOpenAIModelConfiguration
-
-        **Usage**
-
-        .. code-block:: python
-
-            eval_fn = RelevanceEvaluator(model_config)
-            result = eval_fn(
-                question="What is the capital of Japan?",
-                answer="The capital of Japan is Tokyo.",
-                context="Tokyo is Japan's capital, known for its blend of traditional culture \
-                    and technological advancements.")
-        """
         # TODO: Remove this block once the bug is fixed
         # https://msdata.visualstudio.com/Vienna/_workitems/edit/3151324
         if model_config.api_version is None:
             model_config.api_version = "2024-02-15-preview"
 
-        prompty_model_config = {"configuration": model_config}
+        prompty_model_config = {
+            "configuration": model_config,
+        }
+
+        prompty_model_config.update({"parameters": {"extra_headers": {"x-ms-user-agent": USER_AGENT}}})\
+            if USER_AGENT and isinstance(model_config, AzureOpenAIModelConfiguration) else None
+
         current_dir = os.path.dirname(__file__)
         prompty_path = os.path.join(current_dir, "relevance.prompty")
         self._flow = load_flow(source=prompty_path, model=prompty_model_config)
 
     def __call__(self, *, question: str, answer: str, context: str, **kwargs):
-        """Evaluate relevance.
+        """
+        Evaluate relevance.
 
         :param question: The question to be evaluated.
         :type question: str