+++ src/promptflow-evals/tests/evals/unittests/test_eval_run.py
@@ -43,6 +43,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -74,6 +75,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -107,6 +109,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -185,6 +188,7 @@ class TestEvalRun:
                     "info": {
                         "run_id": str(uuid4()),
                         "experiment_id": str(uuid4()),
+                        "run_name": str(uuid4())
                     }
                 }
             },
@@ -193,6 +197,7 @@ class TestEvalRun:
                     "info": {
                         "run_id": str(uuid4()),
                         "experiment_id": str(uuid4()),
+                        "run_name": str(uuid4())
                     }
                 }
             },
@@ -233,6 +238,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -247,9 +253,10 @@ class TestEvalRun:
             workspace_name="mock",
             ml_client=MagicMock(),
         )
-        assert run.info.run_id == mock_response.json.return_value["run"]["info"]["run_id"]
-        assert run.info.experiment_id == mock_response.json.return_value["run"]["info"]["experiment_id"]
-        assert run.name == run.info.run_id
+        assert run.info.run_id == mock_response.json.return_value['run']['info']['run_id']
+        assert run.info.experiment_id == mock_response.json.return_value[
+            'run']['info']['experiment_id']
+        assert run.info.run_name == mock_response.json.return_value['run']['info']["run_name"]
 
     @patch("promptflow.evals.evaluate._eval_run.requests.Session")
     def test_run_with_name(self, mock_session_cls, token_mock, setup_data):
@@ -261,6 +268,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": 'test'
                 }
             }
         }
@@ -275,10 +283,11 @@ class TestEvalRun:
             workspace_name="mock",
             ml_client=MagicMock(),
         )
-        assert run.info.run_id == mock_response.json.return_value["run"]["info"]["run_id"]
-        assert run.info.experiment_id == mock_response.json.return_value["run"]["info"]["experiment_id"]
-        assert run.name == "test"
-        assert run.name != run.info.run_id
+        assert run.info.run_id == mock_response.json.return_value['run']['info']['run_id']
+        assert run.info.experiment_id == mock_response.json.return_value[
+            'run']['info']['experiment_id']
+        assert run.info.run_name == 'test'
+        assert run.info.run_name != run.info.run_id
 
     @patch("promptflow.evals.evaluate._eval_run.requests.Session")
     def test_get_urls(self, mock_session_cls, token_mock, setup_data):
@@ -290,6 +299,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -332,8 +342,11 @@ class TestEvalRun:
         ), "Wrong Metrics URL"
 
     @pytest.mark.parametrize(
-        "log_function,expected_str",
-        [("log_artifact", "allocate Blob for the artifact"), ("log_metric", "save metrics")],
+        'log_function,expected_str',
+        [
+            ('log_artifact', 'register artifact'),
+            ('log_metric', 'save metrics')
+        ]
     )
     def test_log_artifacts_logs_error(self, token_mock, setup_data, tmp_path, caplog, log_function, expected_str):
         """Test that the error is logged."""
@@ -345,6 +358,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -377,24 +391,27 @@ class TestEvalRun:
             # captured by caplog. Here we will skip this logger to capture logs.
             logger.parent = logging.root
             fn = getattr(run, log_function)
-            if log_function == "log_artifact":
-                kwargs = {"artifact_folder": tmp_path}
+            if log_function == 'log_artifact':
+                with open(os.path.join(tmp_path, EvalRun.EVALUATION_ARTIFACT), 'w') as fp:
+                    fp.write('42')
+                kwargs = {'artifact_folder': tmp_path}
             else:
-                kwargs = {"key": "f1", "value": 0.5}
-            fn(**kwargs)
+                kwargs = {'key': 'f1', 'value': 0.5}
+            with patch('promptflow.evals.evaluate._eval_run.BlobServiceClient', return_value=MagicMock()):
+                fn(**kwargs)
         assert len(caplog.records) == 1
         assert mock_response.text in caplog.records[0].message
         assert "404" in caplog.records[0].message
         assert expected_str in caplog.records[0].message
 
     @pytest.mark.parametrize(
-        "dir_exists,expected_error",
-        [
-            (True, "The path to the artifact is empty."),
-            (False, "The path to the artifact is either not a directory or does not exist."),
-        ],
+        'dir_exists,dir_empty,expected_error', [
+            (True, True, "The path to the artifact is empty."),
+            (False, True, "The path to the artifact is either not a directory or does not exist."),
+            (True, False, "The run results file was not found, skipping artifacts upload.")
+        ]
     )
-    def test_wrong_artifact_path(self, token_mock, setup_data, tmp_path, caplog, dir_exists, expected_error):
+    def test_wrong_artifact_path(self, token_mock, tmp_path, caplog, dir_exists, dir_empty, expected_error, setup_data):
         """Test that if artifact path is empty, or dies not exist we are logging the error."""
         mock_response = MagicMock()
         mock_response.status_code = 200
@@ -403,6 +420,7 @@ class TestEvalRun:
                 "info": {
                     "run_id": str(uuid4()),
                     "experiment_id": str(uuid4()),
+                    "run_name": str(uuid4())
                 }
             }
         }
@@ -428,11 +446,14 @@ class TestEvalRun:
             # captured by caplog. Here we will skip this logger to capture logs.
             logger.parent = logging.root
             artifact_folder = tmp_path if dir_exists else "wrong_path_567"
+            if not dir_empty:
+                with open(os.path.join(tmp_path, "test.txt"), 'w') as fp:
+                    fp.write("42")
             run.log_artifact(artifact_folder)
             assert len(caplog.records) == 1
             assert expected_error in caplog.records[0].message
 
-    def test_log_metrics_and_instance_results_logs_error(self, token_mock, setup_data, caplog):
+    def test_log_metrics_and_instance_results_logs_error(self, token_mock, caplog, setup_data):
         """Test that we are logging the error when there is no trace destination."""
         logger = logging.getLogger(ev_utils.__name__)
         # All loggers, having promptflow. prefix will have "promptflow" logger
@@ -440,7 +461,33 @@ class TestEvalRun:
         # captured by caplog. Here we will skip this logger to capture logs.
         logger.parent = logging.root
         ev_utils._log_metrics_and_instance_results(
-            metrics=None, instance_results=None, trace_destination=None, run=None
+            metrics=None,
+            instance_results=None,
+            trace_destination=None,
+            run=None,
+            evaluation_name=None,
         )
         assert len(caplog.records) == 1
         assert "Unable to log traces as trace destination was not defined." in caplog.records[0].message
+
+    def test_run_broken_if_no_tracking_uri(self, setup_data, caplog):
+        """Test that if no tracking URI is provirded, the run is being marked as broken."""
+        logger = logging.getLogger(ev_utils.__name__)
+        # All loggers, having promptflow. prefix will have "promptflow" logger
+        # as a parent. This logger does not propagate the logs and cannot be
+        # captured by caplog. Here we will skip this logger to capture logs.
+        logger.parent = logging.root
+        run = EvalRun(
+            run_name=None,
+            tracking_uri=None,
+            subscription_id='mock',
+            group_name='mock',
+            workspace_name='mock',
+            ml_client=MagicMock()
+        )
+        assert len(caplog.records) == 1
+        assert "The results will be saved locally, but will not be logged to Azure." in caplog.records[0].message
+        with patch('promptflow.evals.evaluate._eval_run.EvalRun.request_with_retry') as mock_request:
+            run.log_artifact('mock_dir')
+            run.log_metric('foo', 42)
+        mock_request.assert_not_called()