+++ src/promptflow-recording/promptflow/recording/azure/utils.py
@@ -4,15 +4,39 @@
 
 import json
 import os
+import queue
 import re
 from dataclasses import dataclass
-from typing import Dict
+from typing import Any, Dict
 
 import jwt
 from azure.core.credentials import AccessToken
 from vcr.request import Request
 
 from .constants import ENVIRON_TEST_PACKAGE, SanitizedValues
+
+
+PF_REQUEST_REPLACEMENTS = {
+    "start_time": SanitizedValues.START_TIME,
+    "timestamp": SanitizedValues.TIMESTAMP,
+    "startTimeUtc": SanitizedValues.START_UTC,
+    "endTimeUtc": SanitizedValues.END_UTC,
+    "end_time": SanitizedValues.END_TIME,
+    "runId": SanitizedValues.RUN_ID,
+    "RunId": SanitizedValues.RUN_ID,
+    # runDisplayName may be the same as RunID
+    "runDisplayName": SanitizedValues.RUN_ID,
+    "container": SanitizedValues.CONTAINER,
+    "flowDefinitionBlobPath": SanitizedValues.FLOW_DEF,
+    "flowArtifactsRootPath": SanitizedValues.ROOT_PF_PATH,
+    "logFileRelativePath": SanitizedValues.EXEC_LOGS,
+    "dataPath": SanitizedValues.DATA_PATH,
+    "Outputs": SanitizedValues.OUTPUTS,
+    "run_id": SanitizedValues.RUN_UUID,
+    "run_uuid": SanitizedValues.RUN_UUID,
+    "exp_id": SanitizedValues.EXP_UUID,
+    "variantRunId": SanitizedValues.RUN_ID,
+}
 
 
 class FakeTokenCredential:
@@ -140,6 +164,57 @@ def sanitize_upload_hash(value: str) -> str:
         value,
         flags=re.IGNORECASE,
     )
+    value = sanitize_pf_run_ids(value)
+    return value
+
+
+def sanitize_pf_run_ids(value: str) -> str:
+    """
+    Replace Run ID on artifacts
+
+    :param value: The value to be modified.
+    :type value: str
+    :returns: The modified string.
+    """
+    re_pf_exp_id = '.+?_\\w{6,8}'
+    re_pf_run_id = '.+?_\\w{6,8}_\\d{8}_\\d{6}_\\d{6}'
+    # Samnutize promptflow-like Run IDs.
+    for left, right, eol in [
+        ("ExperimentRun/dcid[.]", '', ''),
+        ('promptflow/PromptFlowArtifacts/', '', ''),
+        ('runs/', '/flow.flex.yaml', ''),
+        ('runs/', '', '$'),
+        ('runs/', '/.coverage.', ''),
+        ('BulkRuns/', '', '$'),
+            ('runs/', '/batchsync', '')]:
+        value = re.sub(
+            f"{left}{re_pf_run_id}{eol}{right}",
+            f"{left.replace('[.]', '.')}{SanitizedValues.RUN_ID}{right}",
+            value,
+            flags=re.IGNORECASE,
+        )
+    # Sanitize the UUID-like Run IDs.
+    re_uid = '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
+    value = re.sub(
+        f"experimentids/{re_uid}/runs/{re_uid}",
+        f"experimentids/{SanitizedValues.EXP_UUID}/runs/{SanitizedValues.RUN_UUID}",
+        value,
+        flags=re.IGNORECASE,
+    )
+    # And the same for Promptflow experimentids
+    value = re.sub(
+        f"experimentids/{re_pf_exp_id}/runs/{re_pf_run_id}",
+        f"experimentids/{SanitizedValues.EXP_UUID}/runs/{SanitizedValues.RUN_UUID}",
+        value,
+        flags=re.IGNORECASE,
+    )
+    # Sanitize the name of a coverage file
+    value = re.sub(
+        f"runs/{SanitizedValues.RUN_ID}/[.]coverage[^?]+",
+        f"runs/{SanitizedValues.RUN_ID}/{SanitizedValues.COVERAGE}",
+        value,
+        flags=re.IGNORECASE,
+    )
     return value
 
 
@@ -206,19 +281,15 @@ def sanitize_pfs_request_body(body: str) -> str:
     if os.getenv(ENVIRON_TEST_PACKAGE) == "promptflow-azure":
         return json.dumps(body_dict)
 
-    # Start and end time
-    if "start_time" in body_dict:
-        body_dict["start_time"] = SanitizedValues.START_TIME
-    if "timestamp" in body_dict:
-        body_dict["timestamp"] = SanitizedValues.TIMESTAMP
-    if "end_time" in body_dict:
-        body_dict["end_time"] = SanitizedValues.END_TIME
-    # Promptflow Run ID
-    if "runId" in body_dict:
-        body_dict["runId"] = SanitizedValues.RUN_ID
+    # Go over the promptflow replacements
+    for k, v in PF_REQUEST_REPLACEMENTS.items():
+        if k in body_dict:
+            body_dict[k] = v
+
     # Sanitize telemetry event
     if isinstance(body_dict, list) and "Microsoft.ApplicationInsights.Event" in body:
         body_dict = SanitizedValues.FAKE_APP_INSIGHTS
+
     return json.dumps(body_dict)
 
 
@@ -227,12 +298,53 @@ def sanitize_pfs_response_body(body: str) -> str:
     # BulkRuns/{flowRunId}
     if "studioPortalEndpoint" in body:
         body_dict["studioPortalEndpoint"] = sanitize_azure_workspace_triad(body_dict["studioPortalEndpoint"])
+    if "studioPortalTraceEndpoint" in body:
+        body_dict["studioPortalTraceEndpoint"] = sanitize_azure_workspace_triad(body_dict["studioPortalTraceEndpoint"])
     # TraceSessions
     if "accountEndpoint" in body:
         body_dict["accountEndpoint"] = ""
     if "resourceArmId" in body:
         body_dict["resourceArmId"] = ""
+    # Remove token from the response.
+    if "token" in body_dict and body_dict["token"]:
+        body_dict["token"] = "sanitized_token_value"
+    body_dict = sanitize_name(body_dict)
+    # if "createdBy" in body_dict and "userName" in body_dict["createdBy"]:
+    #     body_dict["createdBy"]["userName"] = "First Last"
+    # if "lastModifiedBy" in body_dict and "userName" in body_dict["lastModifiedBy"]:
+    #     body_dict["lastModifiedBy"]["userName"] = "First Last"
+    if isinstance(body_dict, dict) and isinstance(
+        body_dict.get("run"), dict) and isinstance(body_dict["run"].get("data"), dict) and isinstance(
+            body_dict["run"]["data"].get("tags"), list):
+        for dt_list in body_dict["run"]["data"]["tags"]:
+            if isinstance(dt_list, dict) and dt_list.get("key") == "mlflow.user" and not dt_list.get(
+                    "value", "").startswith("promptflow"):
+                dt_list["value"] = "First Last"
     return json.dumps(body_dict)
+
+
+def sanitize_name(dt_data: Dict[str, Any]) -> Dict[str, Any]:
+    """
+    Sanitize the developer first and last name.
+
+    **Note:** The change happens inline. The dictionary is retured
+    for convenience only.
+    :param dt_data: The dictionary to be modified.
+    :type dt_data: dict.
+    :returns: The modified dictionary.
+    """
+    dict_que = queue.Queue()
+    if isinstance(dt_data, dict):
+        dict_que.put(dt_data)
+    while not dict_que.empty():
+        dt_curr = dict_que.get()
+        for k in dt_curr.keys():
+            if isinstance(dt_curr[k], dict):
+                if k == "createdBy" or k == "lastModifiedBy" and "userName" in dt_curr[k]:
+                    dt_curr[k]["userName"] = "First Last"
+                else:
+                    dict_que.put(dt_curr[k])
+    return dt_data
 
 
 def sanitize_email(value: str) -> str:
@@ -245,7 +357,7 @@ def sanitize_file_share_flow_path(value: str) -> str:
         return value
     start_index = value.index(flow_folder_name)
     flow_name_length = 38  # len("simple_hello_world-01-01-2024-00-00-00")
-    flow_name = value[start_index : start_index + flow_name_length]
+    flow_name = value[start_index: start_index + flow_name_length]
     return value.replace(flow_name, "flow_name")
 
 
@@ -307,4 +419,4 @@ def get_created_flow_name_from_flow_path(flow_path: str) -> str:
     # pytest fixture "created_flow" will create flow on file share with timestamp as suffix
     # we need to extract the flow name from the path
     # flow name is expected to start with "simple_hello_world" and follow with "/flow.dag.yaml"
-    return flow_path[flow_path.index("simple_hello_world") : flow_path.index("/flow.dag.yaml")]
+    return flow_path[flow_path.index("simple_hello_world"): flow_path.index("/flow.dag.yaml")]